Even more error fixes
My clang isn't this strict so I do not get these errors

+++ src/video_core/texture_cache/texture_cache_base.h
@@ -420,7 +420,7 @@ private:
     void TickAsyncDecode();
     u32 GetAdaptiveBatchSize(const PendingUnswizzle& task, size_t queue_size) const;
 
-    u32 GetAdaptiveChunkSize(const PendingUnswizzle &task, size_t queue_size) const;
+    size_t GetAdaptiveChunkSize(const PendingUnswizzle &task, size_t queue_size) const;
 
     void EnforceSamplerBudget();
     void TrimInactiveSamplers(size_t budget);
